+++ src/routes/+page.svelte
@@ -273,6 +273,17 @@
         <Action kind="meta">
             <BigLink
                 smaller
+                to="/design"
+                subtitle={(l) => l.ui.page.landing.link.design}
+                ><Iconified
+                    icon="🎨"
+                    text={(l) => l.ui.page.design.header}
+                />
+            </BigLink>
+        </Action>
+        <Action kind="meta">
+            <BigLink
+                smaller
                 to="/donate"
                 subtitle={(l) => l.ui.page.donate.prompt}
             >